requirements.txt
absl-py
annotated-types
anyio
astunparse
audioread
bcrypt
blinker
cachetools
certifi
cffi
charset-normalizer
click
cloudinary
colorama
contourpy
cycler
decorator
distro
exceptiongroup
Flask
Flask-Cors
flatbuffers
fonttools
fpdf
gast
google-auth
google-auth-oauthlib
google-pasta
groq
grpcio
h11
h5py
httpcore
httpx
idna
importlib_metadata
importlib_resources
itsdangerous
Jinja2
joblib
jsonwebtoken
keras
kiwisolver
lazy_loader
libclang
librosa
llvmlite
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdurl
ml-dtypes
msgpack
namex
numba
numpy
oauthlib
opt_einsum
optree
packaging
pillow
platformdirs
pooch
protobuf
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
Pygments
pyparsing
python-dateutil
python-dotenv
requests
requests-oauthlib
rich
rsa
scikit-learn
scipy
six
sniffio
soundfile
soxr
tensorboard
tensorboard-data-server
tensorflow-estimator
tensorflow-hub
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor
tf_keras
threadpoolctl
typing_extensions
urllib3
Werkzeug
wrapt
zipp
pydub
